Crate pallet_dapp_staking_migration

Purpose of this pallet is to provide multi-stage migration for moving from the old dapps_staking_v2 over to the new dapp_staking_v3.

Approach

Multi-Stage Migration

Since a lot of data has to be cleaned up & migrated, it is necessary to do this in multiple steps. To reduce the risk of something going wrong, nothing is done in mandatory hooks, like on_initialize or on_idle. Instead, a dedicated extrinsic call is introduced, which can be called to move the migration forward. As long as this call moves the migration forward, its cost is refunded to the user. Once migration finishes, the extrinsic call will no longer do anything but won’t refund the call cost either.

Migration Steps

The general approach used when migrating is:

Clean up old pallet’s storage using custom code

Use dedicated dApp staking v3 extrinsic calls for registering dApps & locking funds.

The main benefits of this approach are that we don’t duplicate logic that is already present in dApp staking v3, and that we ensure proper events are emitted for each action which will make indexers happy. No special handling will be required to migrate dApps or locked/staked funds over from the old pallet to the new one, from the indexers perspective.

Final Cleanup

The pallet doesn’t clean after itself, so when it’s removed from the runtime, the old storage should be cleaned up using RemovePallet type.

Re-exports

pub use crate::pallet::DappStakingMigrationHandler;

pub use weights::WeightInfo;

pub use pallet::*;

Modules

pallet

The module that hosts all the FRAME types needed to add this pallet to a runtime.

weights

Autogenerated weights for pallet_dapp_staking_migration