Function pallet_dapp_staking_migration::pallet::dispatchables::migrate

pub fn migrate<T: Config>(weight_limit: Option<Weight>)

Attempt to execute migration steps, consuming up to the specified amount of weight. If no weight is specified, max allowed weight is used.

Regardless of the specified weight limit, it will be clamped between the minimum & maximum allowed values. This means that even if user specifies Weight::zero() as the limit, the call will be charged & executed using the minimum allowed weight.

Warning: Doc-Only

This function is an automatically generated, and is doc-only, uncallable stub. See the real version in Pallet::migrate.